HACS + Folder Move
Moved the API Key to /custom/components which seems to be best practices. Also added Hacs.json for easy updating and install via HACS
