+++ custom_components/ps3/__init__.py
@@ -5,7 +5,7 @@ from datetime import timedelta
 from homeassistant.core import HomeAssistant
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
-from .API.PS3MAPI import PS3MAPIWrapper, SensorError
+from .PS3MAPI import PS3MAPIWrapper, SensorError
 from .const import DOMAIN, PLATFORMS
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
